split bash and ansbile
