+++ aws/ssm/ansible.go
@@ -1,0 +1,7 @@
+package ssm
+
+func (s ssm) RunAnsible() error {
+	s.log.Info("Running ssm ansible command")
+	// TODO: implement
+	return nil
+}
